src/python/requirements.txt
blinker==1.8.2
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==43.0.0
defusedxml==0.7.1
elementpath==4.4.0
Flask==3.0.3
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
lxml==5.2.2
MarkupSafe==2.1.5
pycparser==2.22
pyOpenSSL==24.2.1
pysaml2==7.5.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
six==1.16.0
urllib3==2.2.2
Werkzeug==3.0.3
xmlschema==2.5.1
xmlsec==1.3.14
